+++ src/internals/isPlainObject.js
@@ -1,3 +1,5 @@
+// Public
+
 module.exports = function(testVar) {
   if (!testVar || typeof testVar !== 'object') {
     return false;