+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - toEnum does not muteate input
 - promisify don't accept mutate and always mutate module
 - promisify.all does not mutate input options object
+- removed documentation from code
 
 ## [1.4.0] - 2017-07-19
 ### Added