+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - promisify don't accept mutate and always mutate module
 - promisify.all does not mutate input options object
 - removed documentation from code
+- isPlainPbject
 
 ## [1.4.0] - 2017-07-19
 ### Added